requirements.txt
anyio==3.4.0
asgiref==3.4.1
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
ecdsa==0.17.0
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
PyMySQL==1.0.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
typing_extensions==4.0.0
uvicorn==0.15.0
